added authentication and authorization

+++ src/api/controllers/endpointController.js
@@ -1,12 +1,14 @@
 const asyncHandler = require('express-async-handler')
 const Endpoint = require('../models/endpoiontModel')
+const User = require('../models/userModel')
+
 
 // @desc    Get Endpoint
 // @route   GET /api/endpoint
 // @access  Private
 
 const getEndpoint = asyncHandler(async (req, res) => {
-const endpoints = await Endpoint.find()
+const endpoints = await Endpoint.find({user: req.user.id})
 
     res.status(200).json(endpoints)
 })
@@ -21,7 +23,8 @@ const setEndpoint = asyncHandler(async (req, res) => {
         throw new Error('add text')
     }
     const endpoint = await Endpoint.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(endpoint)
@@ -37,6 +40,20 @@ const updateEndpoint = asyncHandler(async (req, res) => {
     if(!endpoint) {
         res.status(400)
         throw new Error('Not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    //validate access
+    if(endpoint.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('not authorized')
     }
 
     const updatedEndpoint = await Endpoint.findByIdAndUpdate(req.params.id, req.body, {new:true})
@@ -55,6 +72,20 @@ const deleteEndpoint = asyncHandler(async (req, res) => {
     if(!endpoint) {
         res.status(400)
         throw new Error('Not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    //validate access
+    if(endpoint.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('not authorized')
     }
 
     await Endpoint.findByIdAndDelete(req.params.id)
